Allow test case reuse
There were a lot of duplicate tests between
Postgres ODBC and MySQL due to previous case
sensitivity. We now ignore case unless the
column names are quoted, which means the
same queries can be used for both Postgres and MySQL.
This also allows the same queries to be tested
against Postgres ODBC and the native Postgres driver.

Due to differences in results between the dsproxy
and non-dsproxy medical databases, the tests haven't
been fully de-duplicated there.