Make wait timeout lower than test timeout
The wait timeout controls how long to wait when using the waitFor* APIs.
These are used to check that something has appeared on the page, etc.
and give nice error messages when the time elapses without the event
occurring. In comparison when the test timeout elapses, only a generic
message is printed.